requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
distro==1.9.0
firebase-admin==6.7.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.18
ml_dtypes==0.5.0
msgpack==1.1.0
mysql-connector-python==9.0.0
numpy==1.26.2
openai==1.66.3
opencv-contrib-python==4.10.0.84
opencv-python==4.8.1.78
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.2
Pillow==10.1.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycaw==20240210
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.5.2
PyJWT==2.10.1
pyparsing==3.2.0
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-vlc==3.0.21203
pywin32==308
requests==2.32.3
rsa==4.9
scipy==1.14.1
screen_brightness_control==0.24.1
selenium==4.27.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
tkhtmlview==0.3.1
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
